Updating to use indicator release v2.1.0. (#4)

+++ alpaca_bars_repository.go
@@ -36,7 +36,7 @@ func barToSnapshot(bar marketdata.Bar) *asset.Snapshot {
 		High:   bar.High,
 		Low:    bar.Low,
 		Close:  bar.Close,
-		Volume: int64(bar.Volume),
+		Volume: float64(bar.Volume),
 	}
 }
